[Refactor][MoE] Read dtype at forward for the MoE family
All nine MoE constructors stop taking `dtype`. The four leaves build
their kernel on first forward and cache it per dtype; the five composites
had nothing left to forward once the leaves stopped accepting it, so the
parameter simply disappears from them.

`FusedMoEExpertsNopadPersistent3WGFwdOp` anchors its shared dtype
validator on `hidden_states.dtype` — the helper already required output,
`w_gate_up` and `w_down` to agree with it, so the constructor's copy was
the redundant one. `SharedFusedMoE` builds its shared-expert MLP kernel
lazily per dtype; a `Kernel` still takes `dtype`, only the op stops
deciding it.

`build_activation_op` drops its `dtype` parameter: the `FusedGatedOp` it
returns already accepted `None` and deferred the build to forward, so the
MoE pipeline never needed to name a dtype for its activation.

Call sites needed three forms handled beyond a keyword argument: a
positional fourth argument to `MoeUnpermuteFwdOp` that would otherwise
have bound silently to `padded_batch_sum`, and a `dict(...)` of kwargs
unpacked at the call.

+++ benchmarks/ops/bench_moe_grouped_gemm_nopad.py
@@ -49,7 +49,7 @@ def test_moe_grouped_gemm_nopad_bench(
     workload = MoeGroupedGemmNopadWorkload(numel, num_experts, n, k, dtype)
     a, b, true_sizes, true_offsets = workload.gen_inputs()
 
-    op = MoeGroupedGemmNopadFwdOp(numel, num_experts, n, k, dtype=dtype)
+    op = MoeGroupedGemmNopadFwdOp(numel, num_experts, n, k)
     bm = ManifestBenchmark(_OP_NAME, op, workload)
 
     # Warmup: trigger JIT compilation before timed profiling.
